feat: Add "tutor openstack create-template" command
This subcommand enables the user to define their own Magnum cluster
template. This requires specifying a base image name, an external
network to connect to, and flavor names for the control plane
("master") and worker nodes.

It's also possible to specify a specific Kubernetes version, and enable
Magnum's private Docker registry.