fix: Avoid timeout when running the registry command in Docker-in-Docker
When "tutor openstack registry" is invoked in a Docker-in-Docker
(dind) environment, then DOCKER_HOST is set to a "tcp://" URL rather
than a UNIX socket. This results in the registry to not listen on port
5000 of localhost, but rather on port 5000 on the dind service
(typically accessible as http://docker:5000, rather than
http://localhost:5000).

Without this change, this makes the connectivity check retry until it
times out.

Thus, check for whether DOCKER_HOST is set in the environment, and
whether it contains a hostname. If it does, run the connectivity check
against port 5000 of that host, rather than localhost.